+++ routes/drone_routes.js
@@ -3,6 +3,6 @@ const router = express.Router();
 const { listarDrones, criarNovoDrone } = require('../controller/drone_controller');
 
 router.get('/', listarDrones);
-router.post('/', criarNovoDrone); // ✅ novo endpoint
+router.post('/', criarNovoDrone);
 
 module.exports = router;
